refactor: consolidate status types and remove unused modules
- Move all status types into api.rs for better cohesion
- Delete network.rs, performance.rs, consistency.rs, filesystem.rs
- Update mod.rs to remove deleted module references
- Update tests.rs imports to use api module
- Reduce module complexity while maintaining all functionality
