+++ janet-ai-retriever/src/status/mod.rs
@@ -1,20 +1,12 @@
 // Status API modules
 pub mod api;
-pub mod consistency;
 pub mod database;
-pub mod filesystem;
-pub mod network;
-pub mod performance;
 pub mod types;
 
 #[cfg(test)]
 mod tests;
 
 // Re-export all public types for easy access
-pub use api::StatusApi;
-pub use consistency::*;
+pub use api::*;
 pub use database::*;
-pub use filesystem::*;
-pub use network::*;
-pub use performance::*;
 pub use types::*;
